AAAAAAARRRRRRRGH!!!
JavaScript needs an implementation of Decimal. Seriously. (Too many)
hours of wrestling with floating-point arithmetic has _finally_ resulted
in a slider balancer which actually works with > 2 sliders, and with
different step values.

The new implementation seeks to balance all sliders in the group
simultaneously. For example, given three sliders:

    Slider 1: 100
    Slider 2:   0
    Slider 3:   0

Changing the slider 1 value to 80 now increases the value of slider two
and three to 10, instead of just changing one slider to 20. This should
have been easy, but thanks to the vast stupidity which is floating-point
arithmetic, it wasn't. Take one slider whose value is 100, and reduce
its value to 99.9; that means that slider two has to be increased to
0.1, right? No.

    > 100 - 99.9
      0.09999999999999432

So, finally we have Balancer.prototype.snapValue which adjusts floating
point numbers to match the precision of the smallest snap value of the
sliders, which seems to work nicely in all my tests.

Other adjustments include recalculating the flex amount after assigning
to each slider to stop it from bouncing between positive and negative
amounts on each iteration, before finally giving up after 20 iterations,
and a couple of short-circuit `break`s to avoid unnecessary iterations.

So yeah. That was my day. Now a little song:

                 Floating-point arithmetic does suck,
                      It involves mostly luck,
                          And I hate it.

I can't write songs.