Implement UnitConversion.
Handles converting between the internal Quinn value, and the alternate
unit types which may be used when selecting values for a slider (e.g.
number of power plants, GWh, etc).

UnitConversion holds values used to correctly format displayed values
(such as the correct precision for each type), and is used by
InputElementView.

The alternate unit types are currently hard-coded into the JavaScript
InputElement model; I -- or someone (?) -- will have to find a decent way
of implementing this on the server-side. I was originally just going to
marshall a hash into a YAML column, but it seems there should be some
way of sharing them, since many sliders have the same unit types...

InputElementView is still a bit messy in places, but it's getting
there.