Handle double-clicks on slider values.
Sometimes a user would double-click a slider value which shows the value
selector. This would cause the second click to be on the "Update" button
which immediately hid the selector. Now, a click on the "Update" button
within 500ms of the value selector being shown will be ignored.

Fixes GH-269.