Name constraint templates using the key, not ID.
IDs are somewhat volatile and it's easy for databases -- particularly in
development -- to get out of sync when using auto-incrementing IDs.

This change renames the constraint partial templates to use the key of
the constraint which should be the same for everyone, regardless of the
ID of the record.

(Case in point, the new "household_energy_cost" had id:10 when I was
working on it; a fresh export of the production DB followed by a
migration gave it id:9).