Position BusyBox correctly when the window is scrolled.
BusyBox doesn't correctly pplace elements which used fixed position when
the user has scrolled the window down. This leads to the dashboard not
being properly obscured. This isn't a new issue -- it's also present in
the production application.

Fixed so that when the BusyBoxes are shown, we manually iterate through
each one which uses fixed positioning and give it the correct "top" CSS
value.