Show input value on area-disabled sliders. Ref #513.
Quinn is strict about disallowing slider values to be outside the given
minimum and maximum values. We work around that by changing the input
min and max to be equal to the user value.

However, this solution is imperfect since the slider handle always gets
drawn at the far left. ET-Engine patch needed to come up with a better
solution...