IE: Animate sidebar mini-graph in pixels.
Internet Explorer fails to load the page since the sidebar mini-graphs
are animated using a percentage value. When the page loads, jQuery in
IE8 believes the bar's width is NaN, and the animation (and page load)
fails.

I swapped to using pixels which achieves the same effect, with the added
bonus that the bars are now a consistent width (two bars showing 50%
will be the same width), whereas before they were a percentage of the
container width, which varied dependign on the length of the title.