Fix VCR in the controller specs.
VCR expects each example have have a name, which is uses to name the
cassette file for the example. When using the short-form for RSpec
examples (`it { should ... }`), no name is present, and VCR tries to
save the cassette as ".yml".

I'm hopeful that this is the source of the intermittant 502 errors
when running CI builds on Semaphore and that the random failures will
cease.

Only one way to find out...