Disallow saving scenarios before we have an ID.
When starting a new scenario, or loading an existing one, ETModel
fetches the scenario data from ETEngine, before notifying ETModel of the
scenario ID. If the user hit the "save scenario" button after the
initial load from ETEngine was started, but before notification of the
ID had been sent to ETModel, the "api_session_id" would be nil.

This value being nil means that ETModel can no longer save the scenario
to ETEngine since it needs to send the ID.

To fix this, the "save scenario" and "reset scenario" links will be
disabled until the ETModel JavaScripts have obtained the ID from
ETEngine *and* notified ETModel of the ID. Once this has been done, the
links are no longer greyed out and will be clickable.

Closes quintel/etengine#542.