Added correct units to entries in targets_met overview
In combination with changes to several records of the Target model,
all the targets will be displayed with the correct unit.

The display_format attribute of records 4 & 5 need to be set to
'number_with_unit' and their unit attribute needs to be changed
to "&euro; bln" and "&euro;/MWh" respectively.

In addition, the unit attribute of records 7 & 9 should be changed
to "km&sup2;"

Closes #1398