Made tracker line update itself when energy storage chart is locked
Added refresh method such that tracker line is updated when total installed
wind and solar capacity is changed and chart is locked.

Furthermore, added chart image to `assets/images/output_elements` to be shown
in chart picker. Made change to `api/area.rb` to allow for energy storage
section to be excluded from sidebar and chart picker in other markets than
the Netherlands.