Added rake task to send out mailing to users
This rake task is used to send a mail to all users who have indicated that
they like to receive updates regarding the ETM. The text of the mail has to
be stored as a template in `app/views/user_mailer`, either in html or text
format or both. Apart from the name of the template to be used (w/o extension),
one needs to provide te subject of the mail. No e-mails are sent unless force
is set to TRUE.

Example:

`rake send_mail subject="Test mail" template="test_mail" force=TRUE`