Introduced conditional links to information pages
Links to information pages are now conditional and based on the domain in use.
For example, if the ETM is run on the pro.et-model.com domain, then it is
assumed that the information pages can be found at et-model.com. Similarly, if
the ETM domain is beta.pro.et-model.com, then the information pages are to be
found at beta.et-model.com. Finally, if you're using pro.et-model.dev, then the
etcentral application should be linked to et-model.dev.