In-page warning message instead of JavaScript pop-up
Instead of a separate pop-up with a warning message when trying to open a
scenario that was created more than a month ago, a warning message is
displayed at the top of the page. Separate messages are used, dependent on
the scenario being a preset or owned by the user himself.