Loading a scenario waits for the ETE response
Instead of setting the preset scenario ID, and then having the JS
client create a scenario on ETEngine, we instead use ActiveResource
to create a scenario on ETEngine before loading the "play" page.

This means that we only set the new `Current.setting` values once we
have a valid scenario on ETE, and prevents invalid sessions (a session
with no scenario ID) if they user quickly hits "back" after loading a
new scenario.

Closes #1745