Add time curve charts
Adds time curve charts for coal/lignite, crude oil, natural gas,
bio-residues, and uranium. All are represented as a line chart, with a
new "time_curve" chart type added to fulfil this role.

Requires that you run the DB migration to add the charts to the
database, and that you are using quintel/etengine@08601e0 or newer.

Closes quintel/etengine#697