Permit customising country name in the selection drop-down
UK, PL, and ES have had the "test" attribute removed (which had no
effect except to append "(new)" to the name in the selection drop-down).

"(Draft)/(Concept)" has been added to their names in the en_areas and
nl_areas locale files to show that they are early test versions.