Add PDF version of reports
Requires the controller to make an (external) request for the HTML
version of the report so that JS may load and execute, and that XHR
request can be made (which would otherwise be prevented by wicked_pdf
loading the rendered HTML using the file:// protocol).

An optional report_pdf_host option is added to the config file for use
in development environments where you need to run multiple instances in
order to process the request.