When locked charts present, do not render default
If the visitor loads the ETM and one or more charts have been locked, do
not also render the default chart for the page.

Reverts bf53986 and d44609d since this is when it was accidentally
changed to also render the default, and the changes there are no longer
required since /output_elements/batch was added in 7992e5f.

Ref #2356
Ref #2136