Simplify MeritTransformator
Does not need to know about the series, since all it does is slice or
downsample arrays of values. Reduced Downsampler and LoadSlicer to
functions.

Working only with an array of values -- instead of series -- makes
the functions simpler and easier to reuse.

MeritTransformator is now a "static" object since it doesn't need to
retain any state.

Adjusted code style (two-space indentation) for consistency with other
ETM JS files and ESLint.