Allow text links to open zoomed charts
Such a link should have the "open_chart" class, and must have a
"data-chart-key" attribute specifying which chart to open. An optional
"data-chart-format" attribute is supported permitting the user to opt
for a table view instead.

For example:

    <!-- Open the CO2 sankey chart -->
    <a href="#" class="open_chart" data-chart-key="co2_sankey">
      Open CO<sub>2</sub> sankey chart
    </a>

    <!-- Open the security of supply chart as a table -->
    <a
      href="#" class="open_chart"
      data-chart-key="security_of_supply"
      data-chart-format="table"
    >
      Open CO<sub>2</sub> sankey chart
    </a>

Note that not all chart support table format; it is the responsibility
of the user not to create a link to a table view when the chart does not
permit it.

Closes #2358