Add household head demand and supply chart
Requests the total demand and supply for heat, and draws deficit and
buffering series automatically based on which is higher.

Lots of code duplicated from the merit order hourly supply chart, but it
doesn't seem possible to extend that, since it is added to the JS bundle
after the heat demand chart. :/