Sankey scaling compensates for number of nodes
Adjusts the y-axis scaling of sankey charts to reflect the number of
nodes in the largest column, rather than using a magic constant.

Now counts the number of nodes in each column, and adds the node margin
to the height of the tallest column to determine which column is the
largest. The sum of the node margins in the tallest column is then
subtracted from the available chart height when rendering.

Ref #2819