Fix attempting to merit-transform a null series
If data for a series is not present (as may be the case when trying to
render a time-resolved chart when Merit is disabled), some charts would
fail to show the "Merit is not enabled" message.

For now, assume that a missing series is not a fatal error and ignore.
This will likely be fixed by the addition of a "curve" gquery unit.

Ref #2873