Scenario report: Get initial vals for presets
ETEngine sets inputs.json "default" to the preset values rather than
the initial starting values for the region. This is expected for
scenarios based on presets, but prevents showing the original and new
slider values for such scenarios in the report.

Reports for presets will now create a new blank scenario and merge the
default input values.

Closes #2439