Add draw to min / max attributes
Allows a slider to be extended in either direction with custom minimum
and maximum values. The user is not permitted to provide a value which
is outside of the acceptable range of the ETEngine input.

This allows insulation sliders to be limited to a non-zero minimum
value, while still showing a slider which starts at zero.

Ref #2984