Fix opacity of unused series in dynamic demand chart
When a serie is made invisible, it is no longer included in the
list of visibleData. This can cause D3 to reuse a g.serie element
which was previously assigned to another serie. The initial chart
may be:

    1. data.key = serie_one
    2. data.key = serie_two
    3. data.key = serie_three

If serie_one is made invisible, D3 will use that SVG element for
serie_two, and the element for serie_two is used for serie_three.
Since elements are hidden based on their keys, this results in the
old serie_three element still being visible:

    1. data.key = serie_two
    2. data.key = serie_three
    3. data.key = serie_three <- unused serie with old data!

`seriesOpacity` keeps track of series already seen so that they may
be made invisible:

    1. data.key = serie_two
    2. data.key = serie_thre