Dynamic demand chart should skip series with no values
The solar curtailment chart sets that the demand and curtailment legend
items are always visible, even if one of the curves has no values (e.g.
there is no curtailment). This was tripping up rendering of the chart,
as it would try to draw a serie with no items (resulting in an undefined
y position).