Remove existing g.series from chart on redraw
When a slider is (re)set to zero while showing in the chart, the zeroed series will remain in the chart data as there is no new data to overwrite it. Because the series still appears in the legend, its opcatiy is NOT set to 0. Therefore all g.series are removed to ensure a full redraw.