Add localized featured scenario title, description
Adds form for adding localized title and descriptions for featured
scenarios. The form page also has a separate section for unfeaturing the
scenario, with a warning that the title and description will be
reverted.

Whenever a scenario is featured, the custom EN and NL title and
description will be shown, rather than tht title and description set on
the saved scenario itself.