Add feedback option to main navigation
Adds a Stripe-style feedback form allowing visitors to provide quick
feedback, either on the ETM in general, or on the current page.

Some metadata is included to help debug any issues the visitor is
reporting and provide context to their message.

A clone of the active scenario will be created.