Add support for demand and production per sector chart
This re-uses the solar curtailment chart to allow toggling which series
is shown. By default the chart will populate the list of selectable
series based on the series in the chart, optionally grouped according
to their `group` attribute.

The chart may define a custom `config.series_selections` option
overriding this behaviour, which the specified options being used
instead.

Ref #3559