Remove clickable target line legend
Removes the clickable target line legend item on the dynamic demand
chart. The total demand line can no longer be toggled off, but this
doesn't seem like a major loss. A more complete solution would be to
port this chart to D3v6, with our new legend system, but likely it's
worth is with a refresh coming soon.

Closes #3510