Keep the survey dismissed when finished
Improvements to the survey dismissal when the visitor clears their
session.

* Finishing the survey but not dismissing the box would not set any
  dismissUntil in localStorage. If the visitor cleared their session,
  the survey would pop-up again.
* Clicking the "x" close button would set the dismissUntil date to be
  three days in the future, rather than ten years when clicking the main
  "Close survey" button.

This fix will ensure dismissUntil is set to be ten years from now as
soon as the visitor completes the survey, and dismisses permenantly when
clicking the "x" on a completed survey.

They'll still see the "Survey" header nav item if they clear their
session, but won't be nagged again.