Remove related node from solar PV cost sliders
The "related node" link would allow us to show technical properties for
the solar PV node by clicking on a link in the slider description. The
properties for household solar PV would be shown.

However, these sliders affect solar PV in households, buildings, and
solar PV plants. This can result in confusion when users think they are
viewing data for plants or buildings, or when they (more likely) expect
it to provide a summary of all solar PV in the scenario.

The links to the node properties are removed to prevent this confusion.
Users may continue to view the tech properties for PV in households,
buildings, or plants in the relevant section.