Attempt to reload session when failing to refresh token
When failing to refresh the access token, it's possible that the token
cannot be refreshed due to concurrent requests to ETModel, both of which
trigger a request to refresh the expired access token.

When this happens, we sleep for a short period and attempt to reload the
identity session information. If the session appears to have a new token
(from the other concurrent request), we load that information and
continue processing the request as normal. Otherwise, we treat it as an
error and reset the session.